+++ app/Controllers/Recurso_entidad.php
@@ -12,25 +12,36 @@ class Recurso_entidad extends BaseController
     public function nuevo()
     {
         if ($this->session->logueado) {
-            $recurso_entidad = $this->request->getPost();
-            if ($recurso_entidad) {
+            $data = [];
+            $data += $this->fn_sis->get_userdata();
 
-                $url_actual = $recurso_entidad['url_actual'];
-                $data = array(
-                    'id_recurso' => $recurso_entidad['id_recurso'],
-                    'id_entidad' => $recurso_entidad['id_entidad'],
-                    'entidad' => $recurso_entidad['entidad'],
-                );
-                // guardar
-                $this->recurso_entidad_model->save($data);
+            $permisos_usuario = $data['permisos_usuario'];
+            $permisos_requeridos = array(
+                'recurso_entidad.can_edit',
+            );
+            if (has_permission_and($permisos_requeridos, $permisos_usuario)) {
+                $recurso_entidad = $this->request->getPost();
+                if ($recurso_entidad) {
 
-                // registro en bitacora
-                $accion = 'agregó';
-                $entidad = 'recurso_entidad';
-                $valor = $recurso_entidad['id_recurso'] . " " .$recurso_entidad['id_entidad']. " " .$recurso_entidad['entidad'];
-                $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+                    $url_actual = $recurso_entidad['url_actual'];
+                    $data = array(
+                        'id_recurso' => $recurso_entidad['id_recurso'],
+                        'id_entidad' => $recurso_entidad['id_entidad'],
+                        'entidad' => $recurso_entidad['entidad'],
+                    );
+                    // guardar
+                    $this->recurso_entidad_model->save($data);
+
+                    // registro en bitacora
+                    $accion = 'agregó';
+                    $entidad = 'recurso_entidad';
+                    $valor = $recurso_entidad['id_recurso'] . " " .$recurso_entidad['id_entidad']. " " .$recurso_entidad['entidad'];
+                    $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+                }
+                return redirect()->to($url_actual);
+            } else {
+                return redirect()->to(site_url());
             }
-            return redirect()->to($url_actual);
         } else {
             return redirect()->to(site_url("login"));
         }
@@ -39,23 +50,34 @@ class Recurso_entidad extends BaseController
     public function eliminar()
     {
         if ($this->session->logueado) {
-            $recurso_entidad = $this->request->getPost();
-            if ($recurso_entidad) {
+            $data = [];
+            $data += $this->fn_sis->get_userdata();
 
-                $id_recurso_entidad = $recurso_entidad['id_recurso_entidad'];
-                $url_actual = $recurso_entidad['url_actual'];
+            $permisos_usuario = $data['permisos_usuario'];
+            $permisos_requeridos = array(
+                'recurso_entidad.can_edit',
+            );
+            if (has_permission_and($permisos_requeridos, $permisos_usuario)) {
+                $recurso_entidad = $this->request->getPost();
+                if ($recurso_entidad) {
 
-                // registro en bitacora
-                $accion = "eliminó";
-                $entidad = 'recurso_entidad';
-                $valor = $recurso_entidad['id_recurso_entidad'];
-                $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+                    $id_recurso_entidad = $recurso_entidad['id_recurso_entidad'];
+                    $url_actual = $recurso_entidad['url_actual'];
 
-                // eliminado
-                $this->recurso_entidad_model->delete($id_recurso_entidad);
+                    // registro en bitacora
+                    $accion = "eliminó";
+                    $entidad = 'recurso_entidad';
+                    $valor = $recurso_entidad['id_recurso_entidad'];
+                    $this->fn_sis->registro_bitacora($accion, $entidad, $valor);
+
+                    // eliminado
+                    $this->recurso_entidad_model->delete($id_recurso_entidad);
+                }
+
+                return redirect()->to($url_actual);
+            } else {
+                return redirect()->to(site_url());
             }
-
-            return redirect()->to($url_actual);
         } else {
             return redirect()->to(site_url("login"));
         }
